Moving all ui rendering to a ui folder
