+++ src/viewer/rendering/ui/uiDetailWindow.cpp
